+++ src/App.vue
@@ -23,7 +23,7 @@
   });
 </script> -->
 <!-- 预览流程实例 -->
-<!-- <template>
+<template>
   <div id="app">
     <SkillfullInstanceBpmnPreview
       ref="diagramInstanceDomRef"
@@ -38,7 +38,7 @@
   onMounted(() => {
     diagramInstanceDomRef.value.viewInstance(instanceData.data);
   });
-</script> -->
+</script>
 
 <!-- 预览流程模型 -->
 
@@ -57,7 +57,7 @@
 </script> -->
 
 <!-- 模型比较 -->
-<template>
+<!-- <template>
   <div id="app">
     <SkillFullBpmnDiffer ref="diagramDomRef" />
   </div>
@@ -72,4 +72,4 @@
       diagramDomRef.value.openDiagram(bpmn2, 'current');
     });
   });
-</script>
+</script> -->