+++ CHANGELOG.md
@@ -46,5 +46,5 @@ IMPROVEMENTS
 1. Upgrade bpmnjs dependencies
 FEATURES
 1. Add diagram differ
-Notice
-1. Incompatible with previous versions(instance rendering needs to provide wiring information)
+NOTICE
+1. Incompatible with previous versions(instance rendering needs to provide sequenceFlowMap information)